Changes related to Redesign of login page
