+++ client-side/C#/StreetSmartArcGISPro/StreetSmartArcGISPro/AddIns/Pages/Login.cs
@@ -257,7 +257,6 @@ namespace StreetSmartArcGISPro.AddIns.Pages
     {
       _login.Username = _username;
       _login.Password = _password;
-      //_login.IsOAuth = _isOAuth;
 
       Save();
 
@@ -278,7 +277,6 @@ namespace StreetSmartArcGISPro.AddIns.Pages
       _login.Save();
       NotifyPropertyChanged("Credentials");
       NotifyPropertyChanged("Username");
-
     }
 
     #endregion
