+++ client-side/C#/StreetSmartArcGISPro/StreetSmartArcGISPro/AddIns/Views/Converters/BasicAuthFieldsVisibilityConverter.cs
@@ -1,26 +1,22 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Data;
 
 namespace StreetSmartArcGISPro.AddIns.Views.Converters
 {
-  class MarginSplitConverter : IValueConverter
+  public class BasicAuthFieldsVisibilityConverter : IValueConverter
   {
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
       if (value is bool credentials && credentials)
       {
-        return "0,0,0,0";
+        return Visibility.Hidden;
       }
-      return "0,20,0,0";
+      return Visibility.Visible;
     }
 
-    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+    public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
     {
       throw new NotImplementedException();
     }
