+++ client-side/C#/StreetSmartArcGISPro/StreetSmartArcGISPro/AddIns/Views/Converters/GridRowPositionOfUsernameBasedOnLoggingStatusConverter.cs
@@ -6,7 +6,7 @@ namespace StreetSmartArcGISPro.AddIns.Views.Converters
 {
   
 
-  public class IsOAuthToRowConverter : IValueConverter
+  public class GridRowPositionOfUsernameBasedOnLoggingStatusConverter : IValueConverter
   {
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
