+++ client-side/C#/StreetSmartArcGISPro/StreetSmartArcGISPro/AddIns/Views/Converters/OAuthUsernameVisibilityConverter.cs
@@ -1,17 +1,12 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Data;
 using static StreetSmartArcGISPro.Configuration.File.Login;
 
 namespace StreetSmartArcGISPro.AddIns.Views.Converters
 {
-  class OAuthUsernameDisplayConverter: IValueConverter
+  class OAuthUsernameVisibilityConverter: IValueConverter
   {
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
