+++ client-side/C#/StreetSmartArcGISPro/StreetSmartArcGISPro/AddIns/Views/Login.xaml.cs
@@ -1,6 +1,4 @@
-﻿using ArcGIS.Desktop.Internal.Mapping;
-using System;
-using System.Drawing;
+﻿using System;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
