+++ client-side/C#/StreetSmartArcGISPro/StreetSmartArcGISPro/Properties/Resources.fr.resx
@@ -128,7 +128,7 @@
     <value>Attention!</value>
   </data>
   <data name="StreetSmartOnLoginFailedMessage" xml:space="preserve">
-    <value>You need to LogIn to StreetSmart in Settings->StreetSmart->Login</value>
+    <value>You need to LogIn to StreetSmart in Settings-&gt;StreetSmart-&gt;Login</value>
   </data>
   <data name="SystemSearch16" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Images\SystemSearch16.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
@@ -207,23 +207,23 @@ Type de produit:</value>
   <data name="LanguageGeneral" xml:space="preserve">
     <value>Général</value>
   </data>
-	<data name="SigningIn" xml:space="preserve">
+  <data name="SigningIn" xml:space="preserve">
     <value>Signing In</value>
   </data>
-	<data name="SignedIn" xml:space="preserve">
+  <data name="SignedIn" xml:space="preserve">
     <value>Signed In</value>
   </data>
-	<data name="SigningOut" xml:space="preserve">
+  <data name="SigningOut" xml:space="preserve">
     <value>Signing Out</value>
   </data>
-	<data name="SignedOut" xml:space="preserve">
+  <data name="SignedOut" xml:space="preserve">
     <value>Signed Out</value>
   </data>
   <data name="LoginCheck" xml:space="preserve">
     <value>Vérifier</value>
   </data>
-	<data name="LogoutCheck" xml:space="preserve">
-    <value>Log out</value>
+  <data name="LogoutCheck" xml:space="preserve">
+    <value>Se déconnecter</value>
   </data>
   <data name="LoginCheckText" xml:space="preserve">
     <value>Vérifiez la combinaison nom d'utilisateur et mot de passe</value>
@@ -300,7 +300,7 @@ Type de produit:</value>
   <data name="HelpNoPdfViewerInstalledOnYourSystem" xml:space="preserve">
     <value>Il n'y a pas de lecteur de pdf installé sur votre système</value>
   </data>
-	<data name="OpenLocationNotSupportedInScene" xml:space="preserve">
+  <data name="OpenLocationNotSupportedInScene" xml:space="preserve">
     <value>Open Location is not supported in Scene</value>
   </data>
 </root>
